Refactor code and create test folder
Create test folder for test cases, need to put more test cases in in the future

Refactored some code in the src files as well
