+++ src/model_trainer.py
@@ -1,0 +1,10 @@
+# local packages:
+from data_loader import create_dataset
+
+
+def train_model_to_data(model):
+    ds = create_dataset(path="../data/",
+                        shape=(224, 224))
+    print(ds)
+    model.fit(ds, epochs=10)
+    return model
